+++ addon/components/bootstrap-button.js
@@ -1,6 +1,5 @@
 import { or } from '@ember/object/computed';
 import Component from '@ember/component';
-import Ember from 'ember';
 import { task, didCancel } from 'ember-concurrency';
 import layout from '../templates/components/bootstrap-button';
 import PropTypeMixin, { PropTypes } from 'ember-prop-types';
@@ -13,11 +12,6 @@ export default Component.extend(PropTypeMixin, {
   propTypes: { // eslint-disable-line ember/avoid-leaking-state-in-ember-objects
     action: PropTypes.func.isRequired,
     disabled: PropTypes.bool,
-  },
-
-  init() {
-    Ember.Logger.warn('DEPRECATION: bootstrap-button component is being replaced by bootstrap/button in an upcoming version of ember-bootstrap-controls.');
-    this._super(...arguments);
   },
 
   disabledButton: or('asyncTask.isRunning', 'disabled'),