+++ addon/mixins/routes/sorted.js
@@ -1,19 +1,27 @@
 import Mixin from '@ember/object/mixin';
+import { isPresent } from '@ember/utils';
 
 export default Mixin.create({
-  queryParams: {
-    sortCiteria: { refreshModel: true },
-    sortReverse: { refreshModel: true },
+  init() {
+    this.queryParams = Object.assign(
+      {
+        sortCriteria: { refreshModel: true },
+        sortReverse: { refreshModel: true },
+      },
+      this.get('queryParams')
+    );
+
+    return this._super(...arguments);
   },
 
-  sortingParams(params) {
+  sortingParams(params, defaultSortCriteria, defaultSortReverse = false) {
     if (params.sortReverse === true || params.sortReverse === false) {
       // NOTE: Convert T/F to string in order to send across server
       params.sortReverse = params.sortReverse.toString();
     }
     return {
-      criteria: params.sortCriteria ? params.sortCriteria : null,
-      reverse: params.sortReverse ? params.sortReverse : null,
+      criteria: params.sortCriteria ? params.sortCriteria : defaultSortCriteria,
+      reverse: params.sortReverse ? params.sortReverse : defaultSortReverse,
     };
   },
 
@@ -25,4 +33,17 @@ export default Mixin.create({
 
     return this._super(...arguments);
   },
+
+  actions: {
+    onSort(sortCriteria, sortReverse) {
+      const controller = this.get('controller');
+      if(isPresent(controller)) {
+        controller.setProperties({
+          sortCriteria: sortCriteria,
+          sortReverse: sortReverse,
+          pageNumber: 1
+        });
+      }
+    },
+  }
 });